+++ lib/view_component/base.rb
@@ -303,6 +303,7 @@ module ViewComponent
     #     config.view_component.view_component_path = "app/my_components"
     #
     # Defaults to `app/components`.
+    #
     mattr_accessor :view_component_path, instance_writer: false, default: "app/components"
 
     # Parent class for generated components
@@ -310,8 +311,16 @@ module ViewComponent
     #     config.view_component.component_parent_class = "MyBaseComponent"
     #
     # Defaults to "ApplicationComponent" if defined, "ViewComponent::Base" otherwise.
-    mattr_accessor :component_parent_class,
-                   instance_writer: false
+    #
+    mattr_accessor :component_parent_class, instance_writer: false
+
+    # Always generate a component with a sidecar directory:
+    #
+    #     config.view_component.generate_sidecar = true
+    #
+    # Defaults to `false`.
+    #
+    mattr_accessor :generate_sidecar, instance_writer: false, default: false
 
     class << self
       # @private